Updated code

+++ 09-spring-boot-jpa-advanced-mappings/23-jpa-many-to-many-delete-student/src/main/java/com/luv2code/cruddemo/dao/AppDAOImpl.java
@@ -204,9 +204,22 @@ public class AppDAOImpl implements AppDAO {
         // retrieve the student
         Student tempStudent = entityManager.find(Student.class, theId);
 
-        // delete the student
-        entityManager.remove(tempStudent);
+        if (tempStudent != null) {
+
+            // get the courses
+            List<Course> courses = tempStudent.getCourses();
+
+            // break association of all courses for the student
+            for (Course tempCourse : courses) {
+                tempCourse.getStudents().remove(tempStudent);
+            }
+
+            // Now delete the student
+            entityManager.remove(tempStudent);
+        }
     }
+
+
 }
